+++ service-map/src/main/java/fi/nls/oskari/map/analysis/service/AnalysisDbService.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface AnalysisDbService {
 
         public long insertAnalysisRow(final Analysis analysis);
-        public int updateAnalysisCols(final Analysis analysis);
+        public long updateAnalysisCols(final Analysis analysis);
         public Analysis getAnalysisById(long id);
         public List<Analysis> getAnalysisById(List<Long> idList);
         public List<Analysis> getAnalysisByUid(String uid);
@@ -18,5 +18,5 @@ public interface AnalysisDbService {
         public void deleteAnalysisByUid(final String id) throws ServiceException;
         public void deleteAnalysis(final Analysis analysis) throws ServiceException;
         public void mergeAnalysis(final Analysis analysis, final List<Long> ids) throws ServiceException;
-        public int updatePublisherName(final long id, final String uuid, final String name);
+        public long updatePublisherName(final long id, final String uuid, final String name);
 }
